Kotlin conversions

+++ src/main/java/at/hannibal2/skyhanni/features/inventory/chocolatefactory/CFShopPrice.kt
@@ -170,7 +170,7 @@ object CFShopPrice {
             add(Renderable.string("§eChocolate available: §6${ChocolateAmount.CURRENT.formatted}"))
             // TODO add chocolate spend needed for next milestone
             add(Renderable.string("§eChocolate spent: §6${chocolateSpent.addSeparators()}"))
-            add(RenderableUtils.fillTable(table, padding = 5, itemScale = config.itemScale))
+            add(RenderableUtils.fillTable(table, padding = 5, itemScale = config.itemScale.toDouble()))
         }
     }
